+++ src/index.ts
@@ -25,6 +25,8 @@ import { OPERATOR as OPERATOR_IN } from './expression/comparison/in'
 import { OPERATOR as OPERATOR_NOT_IN } from './expression/comparison/not-in'
 import { OPERATOR as OPERATOR_PREFIX } from './expression/comparison/prefix'
 import { OPERATOR as OPERATOR_SUFFIX } from './expression/comparison/suffix'
+import { OPERATOR as OPERATOR_OVERLAP } from './expression/comparison/overlap'
+import { OPERATOR as OPERATOR_UNDEFINED } from './expression/predicate/undefined'
 import { OPERATOR as OPERATOR_AND } from './expression/logical/and'
 import { OPERATOR as OPERATOR_OR } from './expression/logical/or'
 import { OPERATOR as OPERATOR_NOR } from './expression/logical/nor'
@@ -41,6 +43,8 @@ export {
   OPERATOR_NOT_IN,
   OPERATOR_PREFIX,
   OPERATOR_SUFFIX,
+  OPERATOR_OVERLAP,
+  OPERATOR_UNDEFINED,
   OPERATOR_AND,
   OPERATOR_OR,
   OPERATOR_NOR,