+++ changelog.md
@@ -1,8 +1,11 @@
 # illogical changelog
 
+## 1.1.4
+* Add Overlap comparison expression.
+
 ## 1.1.3
 * Add predicate expression types.
-* Add UNDEFINED predicate expression.
+* Add Undefined predicate expression.
 
 ## 1.1.2
 * Value operand supports null and undefined as values.